requirements.txt
# Intellectual AI Chat — full install:  pip install -r requirements.txt
# Python 3.11+ recommended (3.12 tested).

fastapi>=0.115,<1
uvicorn[standard]>=0.30,<0.42
python-dotenv>=1.0,<2
pydantic>=2.5,<3

# API + UI
groq>=0.4,<1
pymongo>=4.6,<5
httpx>=0.27,<0.29
requests>=2.31,<3
streamlit>=1.40,<2
